#28 #20 Moved code for favourites database to QACardActivity
