+++ app/src/main/java/org/liberty/android/fantastischmemo/ui/StudyActivity.java
@@ -92,62 +92,18 @@ public class StudyActivity extends QACardActivity {
     private long newCardCount = 0;
     private long newCardLearnedTodayCount = 0;
 
-    /* favourites deck */
-    private String favouritesDbPath = "/sdcard/favourites.db";
-    private AnyMemoDBOpenHelper favouritesDbHelper;
-
     boolean initialized = false;
 
-    @Inject Scheduler scheduler;
+    @Inject
+    Scheduler scheduler;
 
-    @Inject DictionaryUtil dictionaryUtil;
+    @Inject
+    DictionaryUtil dictionaryUtil;
 
-    @Inject ShareUtil shareUtil;
+    @Inject
+    ShareUtil shareUtil;
 
     private GradeButtonsFragment gradeButtonsFragment;
-
-    public AnyMemoDBOpenHelper getFavouritesDbHelper(){
-        return favouritesDbHelper;
-    }
-
-    public void favouriteCard(Card card){
-        if(favouritesDbHelper == null) {
-            favouritesDbHelper = AnyMemoDBOpenHelperManager.getHelper(getApplicationContext(), favouritesDbPath);
-        }
-
-        favouritesDbHelper.getCardDao().createCard(card);
-    }
-
-    public void unfavouriteCard(Card card){
-        if(favouritesDbHelper == null) {
-            favouritesDbHelper = AnyMemoDBOpenHelperManager.getHelper(getApplicationContext(), favouritesDbPath);
-        }
-
-        favouritesDbHelper.getCardDao().delete(card);
-    }
-
-    public List<Card> getAllFavourites(){
-        if(favouritesDbHelper == null) {
-            favouritesDbHelper = AnyMemoDBOpenHelperManager.getHelper(getApplicationContext(), favouritesDbPath);
-        }
-
-        return favouritesDbHelper.getCardDao().queryForAll();
-    }
-
-    public void emptyFavourtiesDeck(){
-        if(favouritesDbHelper == null) {
-            favouritesDbHelper = AnyMemoDBOpenHelperManager.getHelper(getApplicationContext(), favouritesDbPath);
-        }
-
-        List<Card> favourites = getAllFavourites();
-
-        try {
-            favouritesDbHelper.getCardDao().delete(favourites);
-        } catch (SQLException e) {
-            System.out.println("An error occured while attempting to empty the favourites database.");
-            e.printStackTrace();
-        }
-    }
 
     @Override
     public void onCreate(Bundle savedInstanceState){
